Migrate to 2022 and making building easier.
Detours, IoDeviceTracking, InsightSheath and the gui test app/bed work with Visual Studio 2022.  Macros copy static libraries and the InsightSheath / InsightAPI dll to relevate folders in the soluture directory. Note that the release build copies the release version and the debug build copies debug version.  It's not foolproof yet. One may need to manual build detours first before the rest.  Some of the example/test apps are not finished yet and errors will happen on build.