ProcessMemoryCount_x64 marhal appears fixed for now.
Problem appears to be solved for now at least in 64-bit code.  Issue was I failed to relize/remember that the cb is a UINT/DWORD size in BOTH x64 and x86 bit.    C/C++ was not issue.  The Sheath had cb defined as UINT in x86 (correct) and ULONG in x64 (not correct).   Fixed for now as of most recent push to InsightLogger branch